backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.24.3

# Torch with CPU-only version
torch --index-url https://download.pytorch.org/whl/cpu
torch-geometric